feat: refactoring streams
