+++ autoconfigure/src/main/resources/zipkin-server-kafkastore.yml
@@ -6,6 +6,8 @@ zipkin:
       ensure-topics: ${KAFKA_STORE_ENSURE_TOPICS:true}
       store-directory: ${KAFKA_STORE_DIRECTORY:/tmp/zipkin}
       compression-type: ${KAFKA_COMPRESSION_TYPE:NONE}
+      retention-scan-frequency-ms: ${KAFKA_STORE_RETENTION_SCAN_FREQUENCY:86400000}
+      retention-max-age-ms: ${KAFKA_STORE_RETENTION_MAX_AGE:604800000}
       spans-topic: ${KAFKA_STORE_SPANS_TOPIC:zipkin-spans}
       spans-topic-partitions: ${KAFKA_STORE_SPANS_TOPIC_PARTITIONS:1}
       spans-topic-replication-factor: ${KAFKA_STORE_SPANS_TOPIC_REPLICATION_FACTOR:1}
