+++ storage/src/main/java/zipkin2/storage/kafka/streams/serdes/SpansSerde.java
@@ -11,18 +11,29 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin2.storage.kafka.internal.serdes;
+package zipkin2.storage.kafka.streams.serdes;
 
+import java.util.ArrayList;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.serialization.Serializer;
-import zipkin2.DependencyLink;
-import zipkin2.codec.DependencyLinkBytesDecoder;
-import zipkin2.codec.DependencyLinkBytesEncoder;
+import zipkin2.Span;
+import zipkin2.codec.SpanBytesDecoder;
+import zipkin2.codec.SpanBytesEncoder;
 
+import java.util.List;
 import java.util.Map;
 
-public class DependencyLinkSerde implements Serde<DependencyLink> {
+public class SpansSerde implements Serde<List<Span>> {
+
+  private final SpanBytesDecoder spanBytesDecoder;
+
+  private final SpanBytesEncoder spanBytesEncoder;
+
+  public SpansSerde() {
+    this.spanBytesDecoder = SpanBytesDecoder.PROTO3;
+    this.spanBytesEncoder = SpanBytesEncoder.PROTO3;
+  }
 
   @Override
   public void configure(Map<String, ?> configs, boolean isKey) {
@@ -35,17 +46,16 @@ public class DependencyLinkSerde implements Serde<DependencyLink> {
   }
 
   @Override
-  public Serializer<DependencyLink> serializer() {
-    return new DependencyLinkSerializer();
+  public Serializer<List<Span>> serializer() {
+    return new SpansSerializer();
   }
 
   @Override
-  public Deserializer<DependencyLink> deserializer() {
-    return new DependencyLinkDeserializer();
+  public Deserializer<List<Span>> deserializer() {
+    return new SpansDeserializer();
   }
 
-  public static class DependencyLinkDeserializer
-      implements Deserializer<DependencyLink> {
+  private class SpansSerializer implements Serializer<List<Span>> {
 
     @Override
     public void configure(Map<String, ?> configs, boolean isKey) {
@@ -53,11 +63,9 @@ public class DependencyLinkSerde implements Serde<DependencyLink> {
     }
 
     @Override
-    public DependencyLink deserialize(String topic, byte[] data) {
-      if (data == null) {
-        return null;
-      }
-      return DependencyLinkBytesDecoder.JSON_V1.decodeOne(data);
+    public byte[] serialize(String topic, List<Span> data) {
+      if (data == null) return null;
+      return spanBytesEncoder.encodeList(data);
     }
 
     @Override
@@ -66,7 +74,7 @@ public class DependencyLinkSerde implements Serde<DependencyLink> {
     }
   }
 
-  public static class DependencyLinkSerializer implements Serializer<DependencyLink> {
+  private class SpansDeserializer implements Deserializer<List<Span>> {
 
     @Override
     public void configure(Map<String, ?> configs, boolean isKey) {
@@ -74,11 +82,9 @@ public class DependencyLinkSerde implements Serde<DependencyLink> {
     }
 
     @Override
-    public byte[] serialize(String topic, DependencyLink data) {
-      if (data == null) {
-        return null;
-      }
-      return DependencyLinkBytesEncoder.JSON_V1.encode(data);
+    public List<Span> deserialize(String topic, byte[] data) {
+      if (data == null) return new ArrayList<>();
+      return spanBytesDecoder.decodeList(data);
     }
 
     @Override
