+++ storage/src/main/java/zipkin2/storage/kafka/streams/stores/IndexStoreType.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin2.storage.kafka.internal.stores;
+package zipkin2.storage.kafka.streams.stores;
 
 import org.apache.kafka.streams.processor.StateStore;
 import org.apache.kafka.streams.state.QueryableStoreType;
